removing .npmignore to make npm install work for this repository
The problem with .npmignore right now is that it excludes the src/ directory
not only for the package published on npm, but also when the user specifies
this exact Git repository as a dependency in the package.json file of another
project. As a result, the copy of ol-react in the node_modules/ subdir of that
other project will not include the source, so the postinstall script that I
have added in the previous commit will fail. Not having .npmignore in this repo
solves the problem, and I think this is okay because I don't want to publish
this fork on npm anyway